pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "passlib[bcrpyt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "fastapi",
    "uvicorn",
    "sqlalchemy>=2.0.36",
    "passlib[bcrpyt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "psycopg2-binary>=2.9.11",
    "email-validator>=2.3.0",
]
